pyproject.toml
[project]
name = "archdots"
version = "0.1.0"
description = "simple cli tool for managing dotfiles, packages and configuration scripts"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argparse>=1.4.0",
    "decorator>=5.1.1",
    "plyer>=2.1.0",
    "psutil>=6.1.1",
    "pydantic>=2.10.2",
    "pyside6==6.8.1",
    "pyyaml>=6.0.2",
    "rich>=13.9.4",
    "simple-term-menu>=1.6.6",
]

[project.scripts]
dots = "archdots:main"
